+++ mne/viz/tests/test_topomap.py
@@ -22,7 +22,8 @@ from mne.time_frequency.tfr import AverageTFR
 from mne.utils import slow_test
 
 from mne.viz import plot_evoked_topomap, plot_projs_topomap
-from mne.viz.topomap import _check_outlines, _onselect, plot_topomap
+from mne.viz.topomap import (_check_outlines, _onselect, plot_topomap,
+                             _find_peaks)
 
 # Set our plotters to test mode
 import matplotlib
@@ -78,7 +79,7 @@ def test_plot_topomap():
     plt.close('all')
     mask = np.zeros_like(evoked.data, dtype=bool)
     mask[[1, 5], :] = True
-    evoked.plot_topomap(None, ch_type='mag', outlines=None)
+    evoked.plot_topomap(ch_type='mag', outlines=None)
     times = [0.1]
     evoked.plot_topomap(times, ch_type='eeg', res=res, scale=1)
     evoked.plot_topomap(times, ch_type='grad', mask=mask, res=res)
@@ -212,6 +213,19 @@ def test_plot_topomap():
     pos_xywh = np.c_[pos, np.zeros((n_channels, 2))]
     plot_topomap(data, pos_xywh)
     plt.close('all')
+
+    # Test peak finder
+    axes = [plt.subplot(131), plt.subplot(132)]
+    evoked.plot_topomap(times='peaks', axes=axes)
+    plt.close('all')
+    evoked.data = np.zeros(evoked.data.shape)
+    evoked.data[50][1] = 1
+    assert_array_equal(_find_peaks(evoked, 10), evoked.times[1])
+    evoked.data[80][100] = 1
+    assert_array_equal(_find_peaks(evoked, 10), evoked.times[[1, 100]])
+    evoked.data[2][95] = 2
+    assert_array_equal(_find_peaks(evoked, 10), evoked.times[[1, 95]])
+    assert_array_equal(_find_peaks(evoked, 1), evoked.times[95])
 
 
 def test_plot_tfr_topomap():