+++ mne/viz/topomap.py
@@ -1000,7 +1000,7 @@ def plot_tfr_topomap(tfr, tmin=None, tmax=None, fmin=None, fmax=None,
     return fig
 
 
-def plot_evoked_topomap(evoked, times=None, ch_type=None, layout=None,
+def plot_evoked_topomap(evoked, times="auto", ch_type=None, layout=None,
                         vmin=None, vmax=None, cmap='RdBu_r', sensors=True,
                         colorbar=True, scale=None, scale_time=1e3, unit=None,
                         res=64, size=1, cbar_fmt='%3.1f',
@@ -1015,11 +1015,12 @@ def plot_evoked_topomap(evoked, times=None, ch_type=None, layout=None,
     ----------
     evoked : Evoked
         The Evoked object.
-    times : float | array of floats | None.
-        The time point(s) to plot. If None, the number of ``axes`` determines
+    times : float | array of floats | "auto" | "peaks".
+        The time point(s) to plot. If "auto", the number of ``axes`` determines
         the amount of time point(s). If ``axes`` is also None, 10 topographies
         will be shown with a regular time spacing between the first and last
-        time instant.
+        time instant. If "peaks", finds time points automatically by checking
+        for local maxima in global field power.
     ch_type : 'mag' | 'grad' | 'planar1' | 'planar2' | 'eeg' | None
         The channel type to plot. For 'grad', the gradiometers are collected in
         pairs and the RMS for each pair is plotted.
@@ -1134,19 +1135,26 @@ def plot_evoked_topomap(evoked, times=None, ch_type=None, layout=None,
 
     if isinstance(axes, plt.Axes):
         axes = [axes]
-    if times is None:
+
+    if times == "peaks":
+        npeaks = 10 if axes is None else len(axes)
+        times = _find_peaks(evoked, npeaks)
+    elif times == "auto":
         if axes is None:
             times = np.linspace(evoked.times[0], evoked.times[-1], 10)
         else:
             times = np.linspace(evoked.times[0], evoked.times[-1], len(axes))
     elif np.isscalar(times):
         times = [times]
+
     times = np.array(times)
+
     if times.ndim != 1:
         raise ValueError('times must be 1D, got %d dimensions' % times.ndim)
     if len(times) > 20:
         raise RuntimeError('Too many plots requested. Please pass fewer '
                            'than 20 time instants.')
+
     n_times = len(times)
     nax = n_times + bool(colorbar)
     width = size * nax
@@ -1586,3 +1594,22 @@ def _onselect(eclick, erelease, tfr, pos, ch_type, itmin, itmax, ifmin, ifmax,
     fig[0].canvas.draw()
     plt.figure(fig[0].number)
     plt.show()
+
+
+def _find_peaks(evoked, npeaks):
+    """Helper function for finding peaks from evoked data
+    Returns ``npeaks`` biggest peaks as a list of time points.
+    """
+    from scipy.signal import argrelmax
+    gfp = evoked.data.std(axis=0)
+    order = len(evoked.times) // 30
+    if order < 1:
+        order = 1
+    peaks = argrelmax(gfp, order=order, axis=0)[0]
+    if len(peaks) > npeaks:
+        max_indices = np.argsort(gfp[peaks])[-npeaks:]
+        peaks = np.sort(peaks[max_indices])
+    times = evoked.times[peaks]
+    if len(times) == 0:
+        times = [evoked.times[gfp.argmax()]]
+    return times